requirements.txt
playwright==1.39.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.1
google-generativeai==0.3.1
tqdm==4.66.1
SQLAlchemy==2.0.23
notion-client==2.0.0 
openai>=1.0.0
anthropic>=0.18.0